Ignore changes to Gemfile.lock
Since tmpfs is only available on Docker for Linux, we can simply mount
an ignored file over Gemfile.lock. This way, any changes that the
application startup makes to the Gemfile.lock (such as adding wagons)
will not disturb git.

In case we need to add new dependencies, we can still copy back the
updated Gemfile.lock to the core and add the necessary changes to
version control.